requirements.txt
numpy==1.24.2
opencv-python==4.8.0
Pillow==9.4.0